# docs/rentals-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Rental Search API",
    "description": "This API provides access to publicly available data on the Australian rental market in 2026. The API endpoints and their usage are described in detail below."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    }
  },
  "paths": {
...
    "/user/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "Creates a new user account. A request body containing the user to be registered must be sent.",
        "requestBody": {
          "description": "An object containing the email and password of the user to be registered.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "examples": [
                      "mike@gmail.com"
                    ]
                  },
                  "password": {
                    "type": "string",
                    "examples": [
                      "password"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "examples": [
                        "User created"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Request body incomplete, both email and password are required"
                      ]
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "User already exists"
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "examples": [
                    "Too many requests, please try again later."
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "description": "Log in to an existing user account. A request body containing the user credentials must be sent.",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "description": "The credentials of the user to log in.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "examples": [
                      "mike@gmail.com"
                    ]
                  },
                  "password": {
                    "type": "string",
                    "examples": [
                      "password"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Log in successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "examples": [
                        "ajsonwebtoken"
                      ]
                    },
                    "tokenType": {
                      "type": "string",
                      "examples": [
                        "Bearer"
                      ]
                    },
                    "expiresIn": {
                      "type": "number",
                      "examples": [
                        86400
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid log in request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Request body incomplete, both email and password are required"
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Log in failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Incorrect email or password"
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "examples": [
                    "Too many requests, please try again later."
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/user/debugLogin": {
      "post": {
        "description": "Identical to /user/login, except the generated JWT has an expiry time of 1 second. Used to test API response to expired tokens.",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "description": "The credentials of the user to log in.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "examples": [
                      "mike@gmail.com"
                    ]
                  },
                  "password": {
                    "type": "string",
                    "examples": [
                      "password"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login with a 1-second token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "examples": [
                        "ajsonwebtoken"
                      ]
                    },
                    "tokenType": {
                      "type": "string",
                      "examples": [
                        "Bearer"
                      ]
                    },
                    "expiresIn": {
                      "type": "number",
                      "examples": [
                        1
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid log in request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Request body incomplete, both email and password are required"
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Log in failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Incorrect email or password"
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "examples": [
                    "Too many requests, please try again later."
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/user/{email}/profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "description": "Returns profile information. Use the dropdown to see Public vs Private views.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "examples": [
                        "mike@gmail.com"
                      ]
                    },
                    "firstName": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "examples": [
                        "Michael"
                      ]
                    },
                    "lastName": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "examples": [
                        "Jordan"
                      ]
                    },
                    "dob": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "examples": [
                        "1963-02-17"
                      ]
                    },
                    "address": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "examples": [
                        "123 Fake Street, Springfield"
                      ]
                    }
                  }
                },
                "examples": {
                  "authenticatedOwner": {
                    "summary": "Authenticated Owner",
                    "value": {
                      "email": "mike@gmail.com",
                      "firstName": "Michael",
                      "lastName": "Jordan",
                      "dob": "1963-02-17",
                      "address": "123 Fake Street, Springfield"
                    }
                  },
                  "publicView": {
                    "summary": "Non-Owner / Public",
                    "value": {
                      "email": "mike@gmail.com",
                      "firstName": "Michael",
                      "lastName": "Jordan"
                    }
                  },
                  "newlyCreated": {
                    "summary": "Newly Created User",
                    "value": {
                      "email": "notmike@gmail.com",
                      "firstName": null,
                      "lastName": null,
                      "dob": null,
                      "address": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Invalid JWT token"
                      ]
                    }
                  }
                },
                "examples": {
                  "expired": {
                    "summary": "Token Expired",
                    "value": {
                      "error": true,
                      "message": "JWT token has expired"
                    }
                  },
                  "invalid": {
                    "summary": "Token Invalid",
                    "value": {
                      "error": true,
                      "message": "Invalid JWT token"
                    }
                  },
                  "malformed": {
                    "summary": "Header Malformed",
                    "value": {
                      "error": true,
                      "message": "Authorization header is malformed"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "User not found"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "description": "Updates profile for the authenticated owner.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "dob",
                  "address"
                ],
                "properties": {
                  "firstName": {
                    "type": "string",
                    "examples": [
                      "Michael"
                    ]
                  },
                  "lastName": {
                    "type": "string",
                    "examples": [
                      "Jordan"
                    ]
                  },
                  "dob": {
                    "type": "string",
                    "examples": [
                      "1963-02-17"
                    ]
                  },
                  "address": {
                    "type": "string",
                    "examples": [
                      "123 Fake Street, Springfield"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "examples": [
                        "mike@gmail.com"
                      ]
                    },
                    "firstName": {
                      "type": "string",
                      "examples": [
                        "Michael"
                      ]
                    },
                    "lastName": {
                      "type": "string",
                      "examples": [
                        "Jordan"
                      ]
                    },
                    "dob": {
                      "type": "string",
                      "examples": [
                        "1963-02-17"
                      ]
                    },
                    "address": {
                      "type": "string",
                      "examples": [
                        "123 Fake Street, Springfield"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Invalid input: dob must be a date in the past."
                      ]
                    }
                  }
                },
                "examples": {
                  "incomplete": {
                    "summary": "Body Incomplete",
                    "value": {
                      "error": true,
                      "message": "Request body incomplete: firstName, lastName, dob and address are required."
                    }
                  },
                  "invalidTypes": {
                    "summary": "Invalid Name/Address",
                    "value": {
                      "error": true,
                      "message": "Request body invalid: firstName, lastName and address must be strings only."
                    }
                  },
                  "realDate": {
                    "summary": "Invalid Date",
                    "value": {
                      "error": true,
                      "message": "Invalid input: dob must be a real date in format YYYY-MM-DD."
                    }
                  },
                  "futureDate": {
                    "summary": "Future Date",
                    "value": {
                      "error": true,
                      "message": "Invalid input: dob must be a date in the past."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Not your profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Forbidden"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}